Fixing import type errors while consuming the bundled code.
