--- src/Interfaces/.prettierignore
@@ -1,2 +1,0 @@
-# Ignore artifacts:
-dist
