--- src/Interfaces/MarqueeTitleProps.ts
@@ -1,4 +1,0 @@
-export interface MarqueeTitleProps {
-  title: string;
-  interval?: number;
-}
