--- src/Interfaces/SwapFaviconProps.ts
@@ -1,5 +1,0 @@
-import { SwapProps } from "./SwapProps";
-
-export interface SwapFaviconProps extends SwapProps {
-  favicon: string;
-}
