--- src/Interfaces/SwapProps.ts
@@ -1,5 +1,0 @@
-export interface SwapProps {
-  when?: "now" | "onfocus" | "onblur";
-  reset?: "none" | "after" | "onfocus" | "onblur";
-  resetAfterMs?: number;
-}
