--- src/Interfaces/SwapTitleProps.ts
@@ -1,5 +1,0 @@
-import { SwapProps } from "./SwapProps";
-
-export interface SwapTitleProps extends SwapProps {
-  title: string;
-}
