--- src/Interfaces/TabkyJs.ts
@@ -1,6 +1,0 @@
-export interface TabkyJs {
-  originalTitle: string;
-  focusCallbacks: CallableFunction[];
-  blurCallbacks: CallableFunction[];
-	marqueeTitleInterval: ReturnType<typeof setInterval>|null;
-}
